Reset Exterior Cells Only Option
Added requested option in configuration section to disable resetting of any interior cells.
Noticed an issue with the `resetNormalCellsOnRestart` option, as it always deletes exterior cells when true, regardless of if the exterior cell is set as exempt. Best to avoid this toggle for now until I can grasp why it occurs.