Actor Active Spell Stack Limit
This script allows server owners to set active spell effect limits to the actors on their server.
Effects that are added to this scripts table(s) will only allow for one application of the highest magnitude value of said effect to be applied to the actor (with customized maximum magnitudes).

Why might someone want to use this functionality? To maintain some form of relative balance when it comes to spells such as Weakness To <insert element> as players tend to stack multiple spells with said effect, often leading to easy one shots. (i.e., This prevents players from casting 5 different spells with Weakness To Fire 100, then casts a Fireball spell for 500% damage, and instead only allows for 1 instance of Weakness To Fire 100 and a Fireball for 100% extra damage.)